Added bar population script

+++ HCIProject/app/src/main/java/project/hci/hciproject/realm/Bar.java
@@ -9,13 +9,13 @@ import io.realm.annotations.PrimaryKey;
 
 public class Bar extends RealmObject {
     @PrimaryKey private String bar_name;
-    private long latitude;
-    private long longitude;
+    private double latitude;
+    private double longitude;
 
     public Bar() {
     }
 
-    public Bar(String bar_name, long latitude, long longitude) {
+    public Bar(String bar_name, double latitude, double longitude) {
         this.bar_name = bar_name;
         this.latitude = latitude;
         this.longitude = longitude;
@@ -29,19 +29,19 @@ public class Bar extends RealmObject {
         this.bar_name = bar_name;
     }
 
-    public long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(long latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(long longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 }
